refactor: fix compile command
